메모리 증가 이슈 해결, 런타임 중간에 터지는 이슈 해결, 렌더링이슈 발생. 왜안그려지나? 현재 Scene의 렌더링 변수들(루트시그니쳐, CBV, PSO)등 확인중

+++ Client/Client_Prototype/Model_Animation.cpp
@@ -40,7 +40,7 @@ LoadAnimation::LoadAnimation(string filename, float trigger, float skip)
 		now_time = start_time;
 
 		m_fAnimSpeed = (end_time - start_time) / m_pAnim->mChannels[0]->mNumPositionKeys;
-		m_fAnimSpeed = m_pAnim->mTicksPerSecond;
+		m_fAnimSpeed = (float)m_pAnim->mTicksPerSecond;
 		m_fInitSpeed = m_fAnimSpeed;
 	}
